test refactor, renamed feature
