+++ tests/interface/page_features.py
@@ -1,0 +1,3 @@
+# page specific features
+HIDE_REVISION_DATE_ON_PAGE = "revision_date"
+HIDE_REVISION_HISTORY_ON_PAGE = "revision_history"
