+++ tests/interface/theme_features.py
@@ -1,0 +1,4 @@
+# theme features
+SHOW_REVISION_DATE = "revision.date"
+SHOW_REVISION_HISTORY = "revision.history"
+HIDE_SEARCH_BUTTON = "navigation.top.search_button.hide"
