+++ tests/interface/theme_plugins.py
