--- tests/utils/theme_features.py
@@ -1,8 +1,0 @@
-# theme features
-SHOW_REVISION_DATE = "revision.date"
-SHOW_REVISION_HISTORY = "revision.history"
-HIDE_SEARCH_BUTTON = "navigation.top.search.hide"
-
-# page specific features
-HIDE_REVISION_DATE_ON_PAGE = "revision_date"
-HIDE_REVISION_HISTORY_ON_PAGE = "revision_history"
